+++ zeitgitter/commit.py
@@ -129,7 +129,7 @@ def do_commit():
             logging.info("Pushing upstream to %s" % r);
             push_upstream(repo, r, branches)
 
-        if zeitgitter.config.arg.email_address:
+        if zeitgitter.config.arg.mail_address:
             logging.info("cross-timestamping by mail")
             zeitgitter.mail.async_email_timestamp(preserve)
         logging.info("do_commit done")