+++ zeitgitter/tests/20_mail_tests.py
@@ -68,12 +68,11 @@ class MailTests(unittest.TestCase):
             '--own-url', 'https://hagrid.snakeoil',
             '--max-parallel-timeout', '1',
             '--repository', self.tmpdir.name,
-            '--email-address', os.environ['ZEITGITTER_MAILADDRESS'],
+            '--mail-address', os.environ['ZEITGITTER_MAILADDRESS'],
             '--imap-server', os.environ['ZEITGITTER_IMAP_SERVER'],
             '--smtp-server', os.environ['ZEITGITTER_SMTP_SERVER'],
             '--mail-username', os.environ['ZEITGITTER_USERNAME'],
             '--mail-password', os.environ['ZEITGITTER_PASSWORD'],
-            '--email-address', os.environ['ZEITGITTER_MAILADDRESS'],
             # Send test mails to self
             '--external-pgp-timestamper-to', os.environ['ZEITGITTER_MAILADDRESS'],
             '--external-pgp-timestamper-reply', os.environ['ZEITGITTER_MAILADDRESS']